% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_idf.R
\name{calculateIDF}
\alias{calculateIDF}
\title{Calculate Intensity-Duration-Frequency (IDF) models}
\usage{
calculateIDF(data)
}
\arguments{
\item{data}{A data.table object containing imported data.}
}
\value{
A data.table containing IDF models.
}
\description{
This function calculates IDF models based on the input data.
}
\examples{
\dontrun{
  # Assuming `importNetCDFFiles` is correctly implemented and tested
  data <- importNetCDFFiles(path = "/path/to/nc/files")
  calculateIDF(data)
}
}
