% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf_data_extractor.R
\name{importNetCDFFiles}
\alias{importNetCDFFiles}
\title{Import NetCDF (*.nc) files}
\usage{
importNetCDFFiles(path, ids = c(305, 293, 326, 278))
}
\arguments{
\item{path}{Character string. The path to the directory containing NetCDF files.}

\item{ids}{Numeric vector. The cell IDs to extract from each NetCDF file. Default is c(296, 263, 264, 265, 295, 297, 327, 328, 329).}
}
\value{
A data.table containing imported data.
}
\description{
This function imports NetCDF (*.nc) files and returns a data.table containing the imported data.
}
\examples{
\dontrun{
  importNetCDFFiles(path = "/path/to/nc/files")
  importNetCDFFiles(path = "/path/to/nc/files", ids = c(1, 2, 3))
}
}
