% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_visualization.R
\name{plotIDFCurves}
\alias{plotIDFCurves}
\title{Plot IDF Curves}
\usage{
plotIDFCurves(idf_data)
}
\arguments{
\item{idf_data}{A data.table containing IDF models.}
}
\value{
A ggplot object displaying IDF curves.
}
\description{
This function plots IDF curves.
}
\examples{
\dontrun{
  # Assuming `hydroPck_calculate_idf` is correctly implemented and tested
  idf_data <- hydroPck_calculate_idf(data)
  plotIDFCurves(idf_data)
}
}
